finished refactoring PathData
