+++ src/pathpyG/__init__.py
@@ -22,6 +22,8 @@ from pathpyG.core.IndexMap import IndexMap
 from pathpyG.core.HigherOrderIndexMap import HigherOrderIndexMap
 from pathpyG.core.TemporalGraph import TemporalGraph
 from pathpyG.core.PathData import PathData
+from pathpyG.core.WalkData import WalkData
+from pathpyG.core.DAGData import DAGData
 from pathpyG.core.HigherOrderGraph import HigherOrderGraph
 from pathpyG import io
 from pathpyG import nn
