+++ tests/core/test_higher_order_graph.py
@@ -3,12 +3,13 @@ from __future__ import annotations
 from pathpyG.algorithms.temporal import temporal_graph_to_event_dag
 from pathpyG.core.HigherOrderGraph import HigherOrderGraph
 from pathpyG.core.TemporalGraph import TemporalGraph
-from pathpyG.core.PathData import PathData
+from pathpyG.core.DAGData import DAGData
+from pathpyG.core.WalkData import WalkData
 
 
 def test_higher_order_from_temporal_graph(simple_temporal_graph: TemporalGraph):
     dag = temporal_graph_to_event_dag(simple_temporal_graph, delta=5, sparsify=True)
-    paths = PathData.from_temporal_dag(dag, detect_walks=False)
+    paths = DAGData.from_temporal_dag(dag)
 
     g1 = HigherOrderGraph(paths, order=1)
 
@@ -28,19 +29,19 @@ def test_higher_order_from_temporal_graph(simple_temporal_graph: TemporalGraph):
     assert g3.M == 2
     assert g3.data.edge_weight.sum() == 2.0
 
-def test_successors(simple_paths: PathData):
+def test_successors(simple_paths: WalkData):
     g2 = HigherOrderGraph(simple_paths, order=2)
 
     assert set([('C', 'D')]) == set(g2.successors(('A', 'C')))
     assert set([('C', 'E')]) == set(g2.successors(('B', 'C')))
 
-def test_predecessors(simple_paths: PathData):
+def test_predecessors(simple_paths: WalkData):
     g2 = HigherOrderGraph(simple_paths, order=2)
 
     assert set([('A', 'C')]) == set(g2.predecessors(('C', 'D')))
     assert set([('B', 'C')]) == set(g2.predecessors(('C', 'E')))
 
-def test_higher_order_graph(simple_paths: PathData):
+def test_higher_order_graph(simple_paths: WalkData):
     g2 = HigherOrderGraph(simple_paths, order=2)
     assert set(g2.nodes) == set([('A', 'C'), ('C', 'D'), ('B', 'C'), ('C', 'E')])
     assert set(g2.edges) == set([(('A', 'C'), ('C', 'D')), (('B', 'C'), ('C', 'E'))])
